lets generalize exotel code
